:recycle: refactor(rag-service): Small code improvements
