+++ src/rag-service/data_poisoning_attacks/utils/run_data_poisoning_attack.py
@@ -2,17 +2,12 @@ from pathlib import Path
 
 from evaluation.utils.check_connection import check_connection
 from data_poisoning_attacks.utils.ingest_poisoned_embeddings import ingest_poisoned_entries
+from rag_service.constants import RAG_SERVICE_LOCAL_URL, INGESTION_ENDPOINT
 from rag_service.rag_pipeline.utils.read_precomputed_embeddings import (
     get_list_of_embeddings_files,
     read_embeddings_files,
     validate_embeddings_directory
 )
-
-RAG_SERVICE_PORT = 8000
-LOCALHOST = "http://127.0.0.1"
-RAG_SERVICE_LOCAL_URL = LOCALHOST + ":" + str(RAG_SERVICE_PORT) + "/"
-INGESTION_ENDPOINT = "/ingestBatchWithEmbeddingsPrecomputed"
-
 
 def run_data_poisoning_attack(embeddings_dir: Path, logger):
     """ Runs a data poisoning attack on the RAG service with precomputed embeddings from the given Path"""
