+++ src/rag-service/rag_service/rag_pipeline/rag.py
@@ -8,8 +8,8 @@ from llama_index.core.retrievers import VectorIndexRetriever
 from llama_index.core.query_engine import RetrieverQueryEngine
 
 from rag_service.constants import PROVIDER_OLLAMA, PROVIDER_LANGDOCK
-from rag_sevice.rag_pipeline.utils.init_langdock_models import init_langdock_models
-from rag_sevice.rag_pipeline.utils.init_ollama_models import init_ollama_models
+from rag_service.rag_pipeline.utils.init_langdock_models import init_langdock_models
+from rag_service.rag_pipeline.utils.init_ollama_models import init_ollama_models
 from rag_service.config import get_settings
 from logger.logging_config import get_logger
 from rag_service.rag_pipeline.utils.read_precomputed_embeddings import (
