+++ src/rag-service/rag_service/rag_pipeline/utils/read_precomputed_embeddings.py
@@ -24,7 +24,7 @@ def get_list_of_embeddings_files(embeddings_dir: Path) -> List[Path]:
         raise FileNotFoundError(f"Error: Could not find valid files embeddings_part_*.jsonl in {embeddings_dir}")
     return list_of_files
 
-def _read_embeddings_files(embedding_files: list[Path], logger) -> Iterator[Dict[str, Any]]:
+def read_embeddings_files(embedding_files: list[Path], logger) -> Iterator[Dict[str, Any]]:
     """Reads the JSON records from each embedding file"""
     for embedding_file in embedding_files:
         logger.info("Reading %s", embedding_file.name)
@@ -48,7 +48,7 @@ def _get_batch_of_records(records_iter: Iterable[Dict[str, Any]], batch_size: in
 def load_embeddings_into_collection(collection, list_of_embedding_files: list[Path], logger, batch_size: int = 2000) -> int:
     """Loads contents of embeddings files into a Chroma collection in batches."""
     doc_embeddings_added = 0
-    for batch in _get_batch_of_records(_read_embeddings_files(list_of_embedding_files, logger), batch_size):
+    for batch in _get_batch_of_records(read_embeddings_files(list_of_embedding_files, logger), batch_size):
         texts: Documents = [r["text"] for r in batch]
         embeddings: Embeddings = [r["embedding"] for r in batch]
         metadatas: Metadatas = [{"label": r["label"]} for r in batch]
